backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
stripe==11.4.1
openai==1.59.6
cloudinary==1.41.0
redis==5.2.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
asgi-lifespan==2.1.0
